+++ demo/app/components/gwinput-demo.component.ts
@@ -65,10 +65,10 @@ import {Observable} from "rxjs/Observable";
             <div class="col-md-4">
                 <gw-input #gwcontrol
                           [label]="'设备信息'"
-                          [gwClass]="'form-control'"
+                          [gwClass]="'form-control form-control-glowworm'"
                           [btnSize]="'btn-xs'"
                           [enabled]="true"
-                          [closeable]="true"
+                          [closeable]="false"
 
                           [showSelect]="false"
                           [selectData]="selectData"